refactor: refactores user activities architecture
