+++ src/modules/user-activities-answered/infra/http/controllers/user-activities-answered.controller.ts
@@ -9,16 +9,13 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { UserActivitiesAnsweredService } from '../infra/services/user-activities-answered.service';
+import { UserActivitiesAnsweredService } from '../../services/user-activities-answered.service';
 import { Request, Response } from 'express';
-import {
-  AnswerQuestionRequestDTO,
-  AnswerQuestionResponseDTO,
-} from '../domain/dtos/requests/AnswerQuestion.request.dto';
-import { AllExceptionsFilterDTO } from '../../../shared/domain/dtos/errors/AllException.filter.dto';
-import { NotAuthenticatedException } from '../../../shared/domain/errors/NotAuthenticated.exception';
-import { WrongAnswerException } from '../domain/dtos/errors/WrongAnswer.exception';
-import { CommonException } from '../../../shared/domain/errors/Common.exception';
+import { AnswerQuestionRequestDTO, AnswerQuestionResponseDTO } from '../../../domain/dtos/requests/AnswerQuestion.request.dto';
+import { AllExceptionsFilterDTO } from '../../../../../shared/domain/dtos/errors/AllException.filter.dto';
+import { NotAuthenticatedException } from '../../../../../shared/domain/errors/NotAuthenticated.exception';
+import { WrongAnswerException } from '../../../domain/dtos/errors/WrongAnswer.exception';
+import { CommonException } from '../../../../../shared/domain/errors/Common.exception';
 
 @Controller('course')
 @ApiTags('Atividades do Curso')
