--- src/modules/user-courses-concluded/dto/user-courses-concluded.dto.spec.ts
@@ -1,7 +1,0 @@
-import { UserCoursesConcludedDto } from './user-courses-concluded.dto';
-
-describe('UserCoursesConcludedDto', () => {
-  it('should be defined', () => {
-    expect(new UserCoursesConcludedDto()).toBeDefined();
-  });
-});
