--- src/modules/user-courses-concluded/dto/user-courses-concluded.dto.ts
@@ -1,1 +1,0 @@
-export class UserCoursesConcludedDto {}
